changed order of generation to copy over template first
